==> code/06_qiime2_part3.ipynb <==
# Processing marker-gene data in QIIME2, part3

**Environment:** qiime2-2020.11

## How to use this notebook:
1. Activate qiime2 conda environment
   ```
   conda activate qiime2-2020.11 
   ```
    
2. Launch Jupyter notebook
   ```
   jupyter notebook
   ```  

## Hide excessive warnings (optional):
import warnings
warnings.filterwarnings('ignore')

## change working directory to the project root directory
%cd ..

## Filter feature table

!qiime feature-table filter-samples \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-no-control.qza \
  --m-metadata-file data/metadata.tsv \
  --p-where "PCRBatch='1'" \
  --o-filtered-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-digesta-no-control.qza

!qiime feature-table filter-samples \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-no-control.qza \
  --m-metadata-file data/metadata.tsv \
  --p-where "Source IN ('Mucosa', 'Feed', 'Water')" \
  --o-filtered-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-mucosa-feed-water-no-control.qza

## Beta diversity analysis

### Core metric results

#### Digesta samples

!qiime diversity core-metrics \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-digesta-no-control.qza \
  --m-metadata-file data/metadata.tsv \
  --p-sampling-depth 10532 \
  --output-dir data/intermediate/qiime2/asv/core-metrics-results-digesta

#### Mucosa, feed and water samples

!qiime diversity core-metrics \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-mucosa-feed-water-no-control.qza \
  --m-metadata-file data/metadata.tsv \
  --p-sampling-depth 10532 \
  --output-dir data/intermediate/qiime2/asv/core-metrics-results-mucosa-feed-water

### Robust Aitchison PCA 

#### Digesta samples

!qiime deicode rpca \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-digesta-no-control.qza \
  --p-min-feature-count 10 \
  --p-min-sample-count 1000 \
  --output-dir data/intermediate/qiime2/asv/robust-Aitchison-pca-digesta

!qiime emperor biplot \
  --i-biplot data/intermediate/qiime2/asv/robust-Aitchison-pca-digesta/biplot.qza \
  --m-sample-metadata-file data/metadata.tsv \
  --m-feature-metadata-file data/intermediate/qiime2/asv/taxonomy-silva132.qza \
  --o-visualization data/intermediate/qiime2/asv/robust-Aitchison-pca-digesta/biplot.qzv \
  --p-number-of-features 8

#### Mucosa, feed and water samples

!qiime deicode rpca \
  --i-table data/intermediate/qiime2/asv/table-filtered-sepp-inserted-mucosa-feed-water-no-control.qza \
  --p-min-feature-count 10 \
  --p-min-sample-count 1000 \
  --output-dir data/intermediate/qiime2/asv/robust-Aitchison-pca-mucosa-feed-water

!qiime emperor biplot \
  --i-biplot data/intermediate/qiime2/asv/robust-Aitchison-pca-mucosa-feed-water/biplot.qza \
  --m-sample-metadata-file data/metadata.tsv \
  --m-feature-metadata-file data/intermediate/qiime2/asv/taxonomy-silva132.qza \
  --o-visualization data/intermediate/qiime2/asv/robust-Aitchison-pca-mucosa-feed-water/biplot.qzv \
  --p-number-of-features 8
