==> code/09_alpha_diversity.Rmd <==
---
title: "Alpha diversity"
author: "Yanxian Li<br><small>Department of Paraclinical Sciences<br>Faculty of Veterinary Medicine<br>Norwegian University of Life Sciences</small>"
date: "<small>`r Sys.Date()`</small>"
output:
  html_document:
    toc: yes
    toc_depth: 3
    number_sections: true
    toc_float: 
      collapsed: true
    code_folding: show
    theme: cerulean
    self_contained: true
  pdf_document: 
    latex_engine: xelatex
  word_document: default  
---

```{r style, echo = FALSE, message = FALSE, warning = FALSE}
require(knitr)
options(width = 120)
opts_chunk$set(message = FALSE, error = FALSE, warning = FALSE, cache = FALSE)
```

`r Hmisc::hidingTOC(buttonLabel = "Outline (hide)", tocSide = "right",  buttonSide = "right")`

# Getting ready

Load packages.

```{r}
library(here) # A Simpler Way to Find Your Files, CRAN 
library(tidyverse) # Easily Install and Load the 'Tidyverse', CRAN 
library(ggtext) # Improved Text Rendering Support for 'ggplot2', CRAN
library(cowplot) # Streamlined Plot Theme and Plot Annotations for 'ggplot2', CRAN
library(ggsignif) # Significance Brackets for 'ggplot2', CRAN 
library(afex) # Analysis of Factorial Experiments, CRAN 
library(ggResidpanel) # Panels and Interactive Versions of Diagnostic Plots, CRAN 
library(emmeans) # Estimated Marginal Means, aka Least-Squares Means, CRAN 

# Set seed
set.seed(1910)
```

Load data.

```{r}
# metadata
mtd <- read_tsv(here("data/metadata.tsv"), comment = "#q2") %>%
  rename(SampleID = "#SampleID") %>%
  select(!contains("qPCR"))

# alpha diversity
alph <- readRDS(here("data/intermediate/qiime2R/alpha_diversity.rds"))
```

Tidy data.

```{r}
alph <- map_dfc(alph, ~.x$data) %>% # merge alpha diversity indices
  rownames_to_column("SampleID") %>%
  inner_join(mtd, by = "SampleID") %>%
  # remove technical replicates
  filter(!(IsTechnicalReplicate == "yes" & PCRBatch == 2)) %>%
  rename(
    "Observed ASVs" = observed_features, 
    "Pielou's evenness" = pielou_evenness, 
    "Shannon's index" = shannon_entropy, 
    "Faith's PD" = faith_pd
    ) %>%
  pivot_longer(
    cols = "Pielou's evenness":"Shannon's index", 
    names_to = "alph_indx", 
    values_to = "value"
    ) %>%
  mutate(
    FishID = as.character(FishID), 
    Diet = factor(Diet, levels = c("REF", "IM")),
    Segment = factor(Segment, levels = c("PI", "DI")),
    Source = factor(Source, levels = c("Digesta", "Mucosa", "Feed", "Water")),
    alph_indx = factor(
      alph_indx, 
      levels = c("Observed ASVs", "Pielou's evenness", "Shannon's index", "Faith's PD")
      )
    )
```

# Plot alpha diversity

```{r, fig.width=6, fig.height=7}
ggplot(alph, aes(x = Diet, y = value)) +
  geom_point(aes(group = Segment), alpha = 0.3, position = position_dodge(0.5)) +
  stat_summary(
    data = filter(alph, !Source %in% c("Feed", "Water")),
    aes(color = Segment, group = Segment), 
    fun.data = "mean_sdl", 
    fun.args = list(mult = 1), 
    geom = "pointrange",  
    size = 0.5,
    position = position_dodge(0.5)
    ) +
  geom_line(
    data = filter(alph, !Source %in% c("Feed", "Water")) %>%
      group_by(alph_indx, Diet, Source, Segment) %>% 
      summarise(mean = mean(value)),
    aes(y = mean, color = Segment, group = Segment, linetype = Segment), 
    size = 0.5, 
    position = position_dodge(0.5)
    ) + 
  facet_grid(alph_indx ~ Source, scales = "free_y") +
  scale_color_brewer(palette = "Dark2") +
  theme_bw(base_size = 12) +
  theme(legend.position = "top", strip.text.y = element_text(angle = 0))
```

# Compare alpha diversity

## Digesta

### Fit linear mixed effects models

Here we use the `mixed` function from the *afex* package to fit linear mixed effects models. Below we set up a maximal model based on the experimental design. The *p*-value of fixed effects is calculated using the Kenward-Roger approximation, which provides the best control against anti-conservative results.

```{r, message=TRUE}
# filter data
alph_dgs <- filter(alph,  Source == "Digesta") %>%
  droplevels()

# split data
alph_dgs_spl <- split(alph_dgs, alph_dgs$alph_indx)

# fit model
lme_dgs <- lapply(alph_dgs_spl, function(x) {
  mixed(
    value ~ Diet * Segment + (1|FishID) + (1|Tank), 
    data = x, 
    method = "KR"
    )
  }
)

lme_dgs
```

### Model diagnostics

Here we use the *ggResidpanel* package to produce a panel of plots for residual diagnostics.

```{r, results='hide'}
lapply(
  seq_along(lme_dgs), 
  function(x) 
  {
  # extract titles 
  main <- ggdraw() + draw_label(names(lme_dgs)[x], fontface='bold')
  
  # make residual diagnostic plots
  resid_panel <- resid_panel(lme_dgs[[x]]$full_model, plots = "all", qqbands = TRUE)
  
  # assemble plots
  plot_grid(main, resid_panel, ncol = 1, rel_heights = c(1, 10))
  }
)
```

Everything looks OK. We proceed to post-hoc tests.

### Post-hoc tests

When the interaction between main factors is significant, marginal means averaged over the levels of other covariates in the model can be misleading. A good practice is to do conditional contrasts.

```{r}
# filter alpha-diversity indices showing significant interactions between the main effects
intr_dgs <- lapply(lme_dgs, function(x) {
  ifelse(x$anova_table["Diet:Segment", "Pr(>F)"] < 0.05, TRUE, FALSE)
  }
)

# conditional contrasts
cc_dgs <- mapply(function(x, y){
  # compute estimated marginal means 
  emm <- emmeans(x, ~ Diet * Segment)
  # make conditional contrasts if the interaction is significant
  if (y == TRUE){
    contrast(
      emm, method = "consec", simple = "each", combine = TRUE, adjust = "holm"
      ) %>%
    summary()
  }
},
  x = lme_dgs, 
  y = intr_dgs
)

cc_dgs
```

## Mucosa

### Fit linear mixed effects models

```{r, message=TRUE}
# filter data
 alph_mcs <- filter(alph,  Source == "Mucosa") %>%
  droplevels()

# split data
 alph_mcs_spl <- split(alph_mcs,  alph_mcs$alph_indx)

# fit model
lme_mcs <- lapply(alph_mcs_spl, function(x) {
  mixed(
    log1p(value) ~ Diet * Segment + (1|FishID) + (1|Tank), 
    data = x, 
    method = "KR"
    )
  }
)
```

The fitted models are singular. Let's look at the estimated random effects.

```{r}
lapply(lme_mcs, function(x) summary(x)$varcor) 
```

The standard deviation of random effect `FishID` is zero for all the alpha indices. We update the models by removing the random effect `FishID`.

```{r, message=TRUE}
lme_mcs1 <- lapply(alph_mcs_spl, function(x) {
  mixed(
    log1p(value) ~ Diet * Segment + (1|Tank), data = x, method = "KR"
    )
  }
)

lme_mcs1
```

The updated models are no longer singular. Robust results should hold in both maximal and reduced model. Let's compare the pattern of significant and non-significant effects before and after removing the random effect `FishID`.

```{r}
map2_dfr(
  lme_mcs, 
  lme_mcs1, 
  ~left_join(nice(.x), nice(.y), by = "Effect", suffix = c("_max", "_reduced")),
  .id = "alpha_diversity"
  )
```

The pattern of significant and non-significant effect is the same for both models. We proceed to model diagnostics.

### Model diagnostics

```{r, results='hide'}
lapply(
  seq_along(lme_mcs1), 
  function(x) 
  {
  # extract titles 
  main <- ggdraw() + draw_label(names(lme_mcs1)[x], fontface='bold')
  
  # make residual diagnostic plots
  resid_panel <- resid_panel(lme_mcs1[[x]]$full_model, plots = "all", qqbands = TRUE)
  
  # assemble plots
  plot_grid(main, resid_panel, ncol = 1, rel_heights = c(1, 10))
  }
)
```

Everything looks OK. We proceed to post-hoc tests.

### Post-hoc tests

Conditional contrasts.

```{r}
# filter alpha-diversity indices showing significant interactions between the main effects
intr_mcs <- lapply(lme_mcs1, function(x) {
  ifelse(x$anova_table["Diet:Segment", "Pr(>F)"] < 0.05, TRUE, FALSE)
  }
)

# Conditional contrasts
cc_mcs <- mapply(function(x, y){
  # Compute estimated marginal means 
  emm <- emmeans(x, ~ Diet * Segment)

  # Make conditional contrasts if the interaction is significant
  if (y == TRUE){
    contrast(
      emm, method = "consec", simple = "each", combine = TRUE, adjust = "holm"
      ) %>%
    summary()
  }
},
  x = lme_mcs1, 
  y = intr_mcs
)

cc_mcs
```

## Water vs mucosa

Aggregate alpha diversity of mucosa samples in each tank.

```{r}
alph_tm <- filter(alph, Source %in% c("Water", "Mucosa")) %>%
  group_by(alph_indx, Source, Segment, Tank) %>%
  summarise(value = mean(value)) %>%
  mutate(
    SampleType = case_when(
      Segment == "PI" ~ "PIM",
      Segment == "DI" ~ "DIM",
      TRUE ~ "Water")
    ) %>%
  ungroup() %>%
  select(-Segment, -Source)
```

Paired t-test.

```{r}
t_test <- alph_tm %>%
  pivot_wider(names_from = SampleType, values_from = value) %>%
  pivot_longer(PIM:DIM, names_to = "SampleType", values_to = "value") %>%
  group_by(alph_indx, SampleType) %>%
  nest() %>%
  mutate(
    t_test = map(data, ~t.test(.x$Water, .x$value, paired = TRUE, alternative = "two.sided")),
    p_raw = map_dbl(t_test, ~pluck(.x, "p.value"))
    ) %>%
  group_by(alph_indx) %>%
  nest() %>%
  mutate(p_adj = map(data, ~p.adjust(.x$p_raw, method = "holm"))) %>%
  unnest(cols = c(data, p_adj)) %>%
  mutate(
    p_raw = formatC(p_raw, format = "f", digits = 3),
    p_adj = formatC(p_adj, format = "f", digits = 3),
    comparison = paste0("Water", "-", SampleType)
    ) %>%
  select(alph_indx, comparison, p_raw, p_adj)

t_test
```

# Figures

## Figure 3

Gather *p*-values of main effects and their interaction.

```{r}
# digesta samples 
p_main_dgs <- lme_dgs %>% 
  map_dfr(~rownames_to_column(.x$anova_table, "term"), .id = "alph_indx") %>%
  mutate(Source = "Digesta")

# mucosa samples 
p_main_mcs <- lme_mcs1 %>% 
  map_dfr(~rownames_to_column(.x$anova_table, "term"), .id = "alph_indx") %>%
  mutate(Source = "Mucosa")

# merge and tidy data for main effect p value annotation
p_main <- bind_rows(p_main_dgs, p_main_mcs) %>%
  select(-c(`num Df`:F)) %>%
  pivot_wider(names_from = term, values_from = "Pr(>F)") %>%
  mutate_if(is.numeric, ~formatC(.x, format = "f", digits = 3)) %>% # format p value
  mutate(
    Source =  factor(Source, levels = c("Digesta", "Mucosa")),
    # format text using markdown (ggtext)
    lab_main = paste0(
      paste0("Diet: *p* = ", Diet), # *p*, italic
     "<br>", # new line
      paste0("Segment: *p* = ", Segment), 
      "<br>",
      paste0("Interaction: *p* = ", `Diet:Segment`)
      ),
    lab_main = gsub("= 0.000", "< 0.001", lab_main)
    )
```

Gather *p*-values of post-hoc tests.

```{r}
# digesta samples 
p_cc_dgs <- bind_rows(cc_dgs, .id = "alph_indx") %>%
  mutate(Source = "Digesta")

# mucosa samples 
p_cc_mcs <- bind_rows(cc_mcs, .id = "alph_indx") %>%
  mutate(Source = "Mucosa")

# merge and tidy data for post-hoc test p value annotation
p_cc <- bind_rows(p_cc_dgs, p_cc_mcs) %>%
  mutate(Source =  factor(Source, levels = c("Digesta", "Mucosa"))) %>%
  # the "end" comes before "start" because we need to use REF/PI as the starting point
  separate(contrast, sep = " - ", c("end", "start")) %>%
  select(alph_indx, Source, Diet, Segment, start, end, p.value) %>%
  rename(p = p.value) %>%
  arrange(
    match(alph_indx, c("Observed ASVs", "Pielou's evenness", "Shannon's index", "Faith's PD")),
    match(start, c("PI", "DI", ".")),
    match(end, c("REF", "IM", "."))
    ) %>%
  mutate(
    # convert p values to corresponding asterisks
    p = ifelse(p < 0.001, "***", p),
    p = ifelse(p >= 0.001 & p < 0.01, "**", p),
    p = ifelse(p >= 0.01 & p < 0.05, "*", p),
    p = ifelse(p >= 0.05, "NS", p),
    # start and end position of horizontal bar on the x axis 
    start = case_when(
      Diet == "REF" & Segment == "." ~ 0.8,
      Diet == "IM" & Segment == "." ~ 1.8,
      Diet == "." & Segment == "PI" ~ 0.8,
      TRUE ~ 1.2
      ),
    end = case_when(
      Diet == "REF" & Segment == "." ~ 1.2,
      Diet == "IM" & Segment == "." ~ 2.2,
      Diet == "." & Segment == "PI" ~ 1.8,
      TRUE ~ 2.2
      ),
    # position of p value on the x axis 
    p_xpos = (start + end) / 2,
    # position of horizontal bar and p value on the y axis 
    y = c(0.90, 0.91, 0.98, 1.0, 0.97, 1.03, 1.05, 1.09)
  ) 
```

Alpha diversity plots.

```{r}
alph_plots <- lapply(levels(alph$alph_indx), function(x) {
  # filter data
  alph <- filter(alph, alph_indx == x)
  
  # alpha diversity group mean
  alph_mean <- filter(alph, !SampleType %in% c("Feed", "Water")) %>%
    group_by(alph_indx, Diet, Source, Segment) %>% 
    summarise(mean = mean(value))
  
  # p value of main effects
  p_main <- filter(p_main, alph_indx == x)
  
  # p value of conditional contrasts
  p_cc <- filter(p_cc, alph_indx == x)
  
  # make plots
  ggplot(alph, aes(x = Diet, y = value)) +
    geom_point(
      aes(group = Segment), 
      alpha = 0.3, 
      position = position_dodge(0.8)) +
    # add data summary: mean ± sd
    stat_summary(
      data = filter(alph, !SampleType %in% c("Feed", "Water")),
      aes(color = Segment, group = Segment), 
      fun.data = "mean_sdl", 
      fun.args = list(mult = 1), 
      geom = "pointrange",  
      size = 0.5,
      position = position_dodge(0.8)
      ) +
    # use line to show changes in the group means
    geom_line(
      data = alph_mean,
      aes(y = mean, color = Segment, group = Segment, linetype = Segment), 
      size = 0.5, 
      position = position_dodge(0.8)
      ) + 
    # add p values of main effects
    geom_richtext(
      data = p_main,
      aes(x = Inf, y = Inf, label = lab_main), 
      fill = NA, # remove background 
      label.color = NA, # remove outline
      size = 3.5, 
      hjust = 1, 
      vjust = 1.05
      ) +
    # the following two geoms add p values of post-hoc tests
    geom_segment( # add bar
      data = p_cc,
      aes(x = start, xend = end, y = y, yend = y)
      ) +
    geom_text( # add p values
      data = p_cc,
      aes(x = p_xpos, y = y, label = p), 
      size = 3, 
      vjust = -0.1
      ) +
    facet_wrap(~Source, nrow = 1) +
    scale_color_brewer(palette = "Dark2") +
    scale_y_continuous(expand = expansion(mult = c(0.03, 0.4))) +
    labs(x = "", y = x, color = "Segment", linetype = "Segment") +
    theme_bw(base_size = 12) +
    theme(legend.position = "top", strip.text.y = element_blank())
  }
)
```

Assemble plots.

```{r, fig.width=7, fig.height=11}
plot_grid(
  alph_plots[[1]] + 
    theme(axis.text.x = element_blank(), 
          legend.margin = margin(0, 0, 0, 0),
          legend.box.margin = margin(0, 0, -8, 0)),
  alph_plots[[2]] + 
    theme(axis.text.x = element_blank(), 
          legend.position = "none",
          plot.margin = margin(t = -0.7, unit = "cm")),
  alph_plots[[3]] + 
    theme(axis.text.x = element_blank(), 
          legend.position = "none",
          plot.margin = margin(t = -0.5, unit = "cm")),
  alph_plots[[4]] + 
    labs(x = "Diet")+
    theme(legend.position = "none",
          plot.margin = margin(t = -0.5, unit = "cm")),
  ncol = 1, 
  align = "v", 
  rel_heights = c(1.3, 1, 1, 1.1)
  ) 
```

For simplicity, we'll only show Faith's PD and Shannon's index in the Figure 3.

```{r, fig.width=7, fig.height=6}
plot_grid(
  alph_plots[[4]] + 
    theme(
      axis.text.x = element_blank(), 
      legend.margin = margin(0, 0, 0, 0),
      legend.box.margin = margin(0, 0, -8, 0)
      ),
  alph_plots[[3]] +
    labs(x = "Diet") +
    theme(
      legend.position = "none",
      plot.margin = margin(t = -0.5, unit = "cm")
      ),
  ncol = 1, align = "v", rel_heights = c(1.15, 1)) 

# export plot
ggsave(here("result/figure/Figure 3.tiff"), width = 7, height = 6,
       units = "in", dpi = 300, compression = "lzw")
```

## Figure S2

Initial box plot.

```{r}
bp_wt_mcs <- alph_tm %>%
  mutate(SampleType = factor(SampleType, levels = c("Water", "PIM", "DIM"))) %>%
    ggplot(aes(x = SampleType, y = value)) +
    geom_boxplot(aes(color = SampleType)) +
    geom_point(aes(color = SampleType)) +
    facet_wrap(~alph_indx, scales = "free_y", nrow = 1) +
    scale_y_continuous(expand = expansion(mult = c(0.05, 0.1))) +
    scale_color_manual(values = c("#B3B3B3", "#A6CEE3", "#1F78B4")) +
    labs(x = "Sample type") +
    theme_bw(base_size = 12) +
    theme(legend.position = "none")
```

Add *p* values.

```{r, fig.width=8, fig.height=4}
# make a dataframe for p value annotation
p_t.test <- ungroup(t_test) %>%
  separate(comparison, sep = "-", c("start", "end")) %>%
  mutate(p_adj = paste0("italic(p) == ", p_adj),
  y = c(270, 295, 0.95, 0.98, 6.6, 6.9, 41, 44))

# add p values 
bp_wt_mcs <- bp_wt_mcs + 
  geom_signif(
    data = p_t.test,
    aes(xmin = start, xmax = end, annotations = p_adj, y_position = y),
    tip_length = 0.01,
    textsize = 3,
    parse = TRUE,
    manual = T
    )

bp_wt_mcs
```

Figure S2.

```{r, fig.width=4, fig.height=4}
alph_tm %>%
  filter(alph_indx %in% c("Shannon's index", "Faith's PD")) %>%
  mutate(
    alph_indx  = factor(alph_indx, levels = c("Faith's PD", "Shannon's index")),
    SampleType = factor(SampleType, levels = c("Water", "PIM", "DIM"))
    ) %>%
  ggplot(aes(x = SampleType, y = value)) +
    geom_boxplot(aes(color = SampleType)) +
    geom_point(aes(color = SampleType)) +
    geom_signif(
      data = filter(p_t.test, alph_indx %in% c("Shannon's index", "Faith's PD")),
      aes(xmin = start, xmax = end, annotations = p_adj, y_position = y),
      tip_length = 0.01,
      textsize = 3,
      parse = TRUE,
      manual = T
      ) +
    facet_wrap(~alph_indx, scales = "free_y", nrow = 1) +
    scale_y_continuous(expand = expansion(mult = c(0.05, 0.1))) +
    scale_color_manual(values = c("#B3B3B3", "#A6CEE3", "#1F78B4")) +
    labs(x = "") +
    theme_bw(base_size = 12) +
    theme(legend.position = "none")

# export plot
ggsave(here("result/figure/Figure S2.tiff"), width = 4, height = 4,
       units = "in", dpi = 300, compression = "lzw")
```

# Session information

```{r session-info}
sessionInfo()
```
